Fix bug causing games to restart at incorrect times
Previously the game to be played was determined by using the UTC date at
runtime. This was problematic because the UTC date can differ from the
user's local date, causing games to switch over at all times of day
depending on where in the world the user was located. Now the game to be
played is determined by using the local date at runtime and switches
over into the next game when the local date changes.